+++ common/src/main/java/org/broadleafcommerce/common/sitemap/service/SiteMapServiceImpl.java
@@ -20,6 +20,7 @@
 
 package org.broadleafcommerce.common.sitemap.service;
 
+import org.apache.commons.io.FilenameUtils;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.broadleafcommerce.common.config.domain.ModuleConfiguration;
@@ -27,7 +28,6 @@ import org.broadleafcommerce.common.config.service.ModuleConfigurationService;
 import org.broadleafcommerce.common.config.service.type.ModuleConfigurationType;
 import org.broadleafcommerce.common.file.domain.FileWorkArea;
 import org.broadleafcommerce.common.file.service.BroadleafFileService;
-import org.broadleafcommerce.common.file.service.BroadleafFileUtils;
 import org.broadleafcommerce.common.sitemap.domain.SiteMapConfiguration;
 import org.broadleafcommerce.common.sitemap.domain.SiteMapGeneratorConfiguration;
 import org.broadleafcommerce.common.sitemap.exception.SiteMapException;
@@ -58,6 +58,8 @@ import javax.annotation.Resource;
 public class SiteMapServiceImpl implements SiteMapService {
 
     protected static final Log LOG = LogFactory.getLog(SiteMapServiceImpl.class);
+    
+    protected static final String ENCODING_EXTENSION = ".gz";
 
     protected Boolean gzipSiteMapFiles;
 
@@ -111,8 +113,18 @@ public class SiteMapServiceImpl implements SiteMapService {
         }
 
         siteMapBuilder.persistSiteMap();
+
+
+        // Check for GZip
         if (getGzipSiteMapFiles()) {
-            gzipAndDeleteFiles(fileWorkArea, siteMapBuilder.getIndexedFileNames());
+            gzipAndDeleteFiles(fileWorkArea, siteMapBuilder.getIndexedFileNames(), false);
+            List<String> indexFileNames = new ArrayList<String>();
+            for (String fileName: siteMapBuilder.getIndexedFileNames()) {
+                indexFileNames.add(fileName + ENCODING_EXTENSION);
+            }
+            smgr.setSiteMapFilePaths(indexFileNames);
+        } else {
+            smgr.setSiteMapFilePaths(siteMapBuilder.getIndexedFileNames());
         }
 
 
@@ -130,10 +142,21 @@ public class SiteMapServiceImpl implements SiteMapService {
         }
         File siteMapFile = broadleafFileService.getResource(fileName, getSiteMapTimeoutInMillis());
         if (siteMapFile.exists()) {
+
+            if (getAutoGenerateSiteMapAfterTimeout()) {
+                long lastModified = siteMapFile.lastModified();
+                long now = System.currentTimeMillis();
+                // Create new SiteMap if timeout expired.
+                if ((now - lastModified) > getSiteMapTimeoutInMillis().longValue()) {
+                    generateSiteMap();
+                    siteMapFile = broadleafFileService.getResource(fileName, getSiteMapTimeoutInMillis());
+                }
+            }
             if (LOG.isTraceEnabled()) {
                 LOG.trace("Returning existing SiteMap");
             }
             return siteMapFile;
+
         } else {
             if (getCreateSiteMapIfNotFound()) {
                 if (LOG.isTraceEnabled()) {
@@ -194,17 +217,17 @@ public class SiteMapServiceImpl implements SiteMapService {
     }
 
     /**
-     * Gzip a file and then delete the file
-     * 
-     * @param fileName
+     *
+     * @param fileWorkArea
+     * @param fileNames
      */
-    protected void gzipAndDeleteFiles(FileWorkArea fileWorkArea, List<String> fileNames) {
+    protected void gzipAndDeleteFiles(FileWorkArea fileWorkArea, List<String> fileNames,boolean shouldDeleteOriginal){
         for (String fileName : fileNames) {
             try {
-                String fileNameWithPath = BroadleafFileUtils.buildFilePath(fileWorkArea.getFilePathLocation(), fileName);
+                String fileNameWithPath = FilenameUtils.normalize(fileWorkArea.getFilePathLocation() + File.separator + fileName);
 
                 FileInputStream fis = new FileInputStream(fileNameWithPath);
-                FileOutputStream fos = new FileOutputStream(fileNameWithPath + ".gz");
+                FileOutputStream fos = new FileOutputStream(fileNameWithPath + ENCODING_EXTENSION);
                 GZIPOutputStream gzipOS = new GZIPOutputStream(fos);
                 byte[] buffer = new byte[1024];
                 int len;
@@ -216,14 +239,27 @@ public class SiteMapServiceImpl implements SiteMapService {
                 fos.close();
                 fis.close();
 
-                File originalFile = new File(fileNameWithPath);
-                originalFile.delete();
+                if(shouldDeleteOriginal){
+                    File originalFile = new File(fileNameWithPath);
+                    originalFile.delete();
+                }
+
 
             } catch (IOException e) {
                 e.printStackTrace();
             }
         }
     }
+
+    /**
+     * GZip a file, Then delete it
+     * @param fileWorkArea
+     * @param fileNames
+     */
+    protected void gzipAndDeleteFiles(FileWorkArea fileWorkArea, List<String> fileNames) {
+        gzipAndDeleteFiles(fileWorkArea,fileNames,true);
+    }
+
 
     public List<SiteMapGenerator> getSiteMapGenerators() {
         return siteMapGenerators;
@@ -247,6 +283,10 @@ public class SiteMapServiceImpl implements SiteMapService {
 
     public boolean getCreateSiteMapIfNotFound() {
         return BLCSystemProperty.resolveBooleanSystemProperty("sitemap.createIfNotFound");
+    }
+
+    public boolean getAutoGenerateSiteMapAfterTimeout() {
+        return BLCSystemProperty.resolveBooleanSystemProperty("sitemap.createIfTimeoutExpired",false);
     }
 
     public Long getSiteMapTimeoutInMillis() {
